refactor(studio): split the two files over the 600-line cap
The file size check has been failing on main. It is diff-scoped on a PR but
full-scans packages/studio on push, so two files that crept over the cap were
only ever caught after merge, and every release since has been red:

  TimelineAutomationLane.tsx  674 lines
  StudioRightPanel.tsx        609 lines

Both are pure code moves. No behavior change.

TimelineAutomationLane.tsx keeps the single-lane editor and gives up the
track-level layer: ClipLaneRow, ClipAutomationLanes and
TimelineAutomationLaneSlot move to TimelineAutomationLaneSlot.tsx, which is the
name its test file already used. The dependency runs one way, slot -> lane, so
there is no cycle. 674 -> 499.

StudioRightPanel.tsx gives up its props interface to a sibling .types.ts. That
block is the part that changes least, so moving it keeps the component's own
diffs small; two in-flight branches touch this file and both are based on a
556-line copy of it, so keeping the cut away from the body matters. 609 -> 568.

Verified by running the CI rule's full-scan branch over the 679 tracked
packages/studio source files: no file over 600, exit 0. Studio suite green at
2790 passed across 226 files, and typecheck clean.
